BashBuddy/pyproject.toml
[project]
name = "bashbuddy"
version = "0.1.0"
description = ""
authors = [
    {name = "Jannymo13",email = "jonkonichottil@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click (>=8.3.0,<9.0.0)",
    "google-genai (>=1.41.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "supabase (>=2.21.1,<3.0.0)",
    "questionary (>=2.1.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "bashbuddy", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bashbuddy = "bashbuddy:cli"
bb = "bashbuddy:cli"

